rename app to buzzid
