+++ buzzid/annotate/select_regions_of_interest_interactive.py
