+++ buzzid/detect/models.py
